update: have the component just for the input
